Added a message when no completed blocks are found

+++ app/src/main/java/com/example/scarlet/feature_training_log/presentation/training_log/components/CompletedBlocksSection.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.example.scarlet.R
 import com.example.scarlet.feature_training_log.domain.model.Block
 import com.example.scarlet.feature_training_log.presentation.components.ScarletList
+import com.example.scarlet.feature_training_log.presentation.components.SectionTitle
 import com.example.scarlet.feature_training_log.presentation.destinations.BlockScreenDestination
 import com.example.scarlet.feature_training_log.presentation.training_log.TrainingLogEvent
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
@@ -22,27 +23,40 @@ fun CompletedBlocksSection(
     completedBlocks: List<Block> = emptyList(),
     onEvent: (TrainingLogEvent) -> Unit
 ) {
-    ScarletList(
-        modifier = Modifier.fillMaxWidth(),
-        title = stringResource(R.string.completed_training_blocks),
-        items = completedBlocks,
-        onItemClicked = { block ->
-            navigator.navigate(BlockScreenDestination(block))
-        },
-        onDeleteClicked = { block ->
-            onEvent(TrainingLogEvent.DeleteBlock(block))
+    if (completedBlocks.isNotEmpty()) {
+        ScarletList(
+            modifier = Modifier.fillMaxWidth(),
+            title = stringResource(R.string.completed_training_blocks),
+            items = completedBlocks,
+            onItemClicked = { block ->
+                navigator.navigate(BlockScreenDestination(block))
+            },
+            onDeleteClicked = { block ->
+                onEvent(TrainingLogEvent.DeleteBlock(block))
+            }
+        ) { block ->
+            Column {
+                Text(
+                    text = block.name,
+                    style = MaterialTheme.typography.bodyLarge
+                )
+                Text(
+                    text = "XX/XX/XXXX - XX/XX/XXXX", /* TODO*/
+                    style = MaterialTheme.typography.bodySmall
+                )
+            }
         }
-    ) { block ->
-        Column {
+    } else {
+        Column(
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            SectionTitle(title = stringResource(R.string.completed_training_blocks))
             Text(
-                text = block.name,
-                style = MaterialTheme.typography.bodyLarge
-            )
-            Text(
-                text = "XX/XX/XXXX - XX/XX/XXXX", /* TODO*/
-                style = MaterialTheme.typography.bodySmall
+                text = stringResource(R.string.no_completed_training_blocks),
+                style = MaterialTheme.typography.bodyMedium
             )
         }
+
     }
 }
